Create README.md
The Desktop Application, a Transportation Management System, was built based on real-world scenarios identified during a meeting with the company manager. It was developed using Windows Forms Application and C# programming language to offer a user-friendly interface and efficient performance. The system integrates with a SQL Server database to provide robust data management capabilities, making it a reliable solution for the transportation management needs of the company.